start with splitting up runtests.jl
